Changing BOOTPROTO configuration
According to RH documentation and initscripts there is no argument "dhcp6" for config parameter 'BOOTPROTO'. The only valid options are none, dhcp (ipv4) and bootp. Any other option is going to be ignored and treated as "none". That's why "dhcp6" "works". I propose this change in order to get compliant with the oficial operating system documentation, linked below.

https://access.redhat.com/documentation/en-US/Red_Hat_Enterprise_Linux/6/html/Deployment_Guide/s1-networkscripts-interfaces.html